Implement RGB canvas framework.
Definitions:
  palette: array of unique 24-bit RGB colors
  LCT: local color table (palette with 1-frame scope)
  GCT: global color table (palette apllied to all frames lacking LCT)
  frame: image array of 8-bit indices into palette entries
  canvas: image array of 24-bit colors

Initially, gifdec only  stored the GCT, 1  LCT and 1 frame.  In order to
implement  the disposal  method  "restore to  previous"  (RTP), a  "back
frame" was added. However, the back frame only stored color indices, and
those indices could be referencing previous LCTs.

With this commit,  gifdec will store GCT,  1 LCT, 1 frame  and 1 canvas.
Every frame that isn't marked for RTP is rendered to the canvas after it
has been used, apllying the current valid palette -- whether it's GCT or
the frame's LCT. This way, the canvas will store pixel colors that could
otherwise be lost with the old approach.